\name{datafile_details}
\alias{datafile_details}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function : tcltk code to display a window asking for details about the layout of the data file.
}
\description{
tcltk code for a windows which allows to specificy some details about the input file.
}
\usage{
datafile_details()
}
\value{
A binary vector indicating whether row names et column names are present in the input file.
}
\author{
Frederic Santos
}
\note{
By default, the two checkboxes for row and column names are set as "true".
}

\seealso{
See alsol the R package tcltk.
}

