\name{KappaGUI-package}
\alias{KappaGUI-package}
\alias{KappaGUI}
\docType{package}
\title{
GUI for Cohen's and Fleiss' Kappa.
}
\description{
Offers a complete and interactive GUI to work out Cohen's and Fleiss' Kappa, based on the irr library.
}
\details{
\tabular{ll}{
Package: \tab KappaGUI\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2017-06-22\cr
License: \tab GPL >=2\cr
}

}
\author{
Frederic Santos

Maintainer: Frederic Santos <frederic.santos@u-bordeaux.fr>
}
\references{
Cohen, J. (1960). A coefficient of agreement for nominal scales. Educational and Psychological Measurement, 20, 37-46.

Cohen, J. (1968). Weighted kappa: Nominal scale agreement with provision for scaled disagreement or partial credit. Psychological Bulletin, 70, 213-220.
}
\keyword{Kappa}
\seealso{
kappa2
}
\examples{
\donttest{StartKappa()}
}
