\name{KappaGUI-package}
\alias{KappaGUI-package}
\alias{KappaGUI}
\docType{package}
\title{
A Shiny user interface for Cohen's and Fleiss' Kappa.
}
\description{
Offers a Shiny user interface to work out Cohen's and Fleiss' Kappa, using the library 'irr'.
}
\details{
\tabular{ll}{
Package: \tab KappaGUI\cr
Type: \tab Package\cr
Version: \tab 2.0.0\cr
Date: \tab 2017-07-06\cr
License: \tab GPL >=2\cr
}

}
\author{
Frederic Santos, <frederic.santos@u-bordeaux.fr>
}
\references{
Cohen, J. (1960). A coefficient of agreement for nominal scales. Educational and Psychological Measurement, 20, 37-46.

Cohen, J. (1968). Weighted kappa: Nominal scale agreement with provision for scaled disagreement or partial credit. Psychological Bulletin, 70, 213-220.
}
\keyword{Kappa}
\seealso{
kappa2
}
\examples{
\dontrun{ StartKappa() }
}
